Viterbi classification
Wrote code for Viterbi algorithm with continuous emissions. It seems to work properly (classification behaves as expected on phantom data), but doesn't work at all on CFD data. Also, number of states as currently defined is way too high for BW to work in a decent amount of time, making it impossible to extend the algorithm correctly. Going forward, a way of trimming down on the number of states before using HMMs is probably requried.